python_learning/pyproject.toml
[tool.poetry]
name = "python_learning"
version = "0.1.0"
description = ""
authors = ["Zoran Topic <zoran1607@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^1.4.1"
requests = "^2.28.0"
python-dotenv = "^0.20.0"
rich = "^12.5.1"
textual = "^0.1.18"
psycopg = "^3.1.6"
openai = "^0.26.4"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
